+++ types/snapshot.go
@@ -74,6 +74,7 @@ const (
 	FeeTrackerSnapshot             SnapshotNamespace = "feestracker"
 	MarketTrackerSnapshot          SnapshotNamespace = "markettracker"
 	ERC20MultiSigTopologySnapshot  SnapshotNamespace = "erc20multisigtopology"
+	PoWSnapshot                    SnapshotNamespace = "pow"
 
 	MaxChunkSize   = 16 * 1000 * 1000 // technically 16 * 1024 * 1024, but you know
 	IdealChunkSize = 10 * 1000 * 1000 // aim for 10MB